+++ src/constants/api_connection.ts
@@ -197,11 +197,11 @@ export enum APIPath {
   LIST_INVOICE_RC2_INPUT = `${apiPrefixRC2}/account_book/:accountBookId/invoice/input`,
   CREATE_INVOICE_RC2_INPUT = `${apiPrefixRC2}/account_book/:accountBookId/invoice/input`,
   GET_INVOICE_RC2_INPUT = `${apiPrefixRC2}/account_book/:accountBookId/invoice/:invoiceId/input`,
-  DELETE_INVOICE_RC2_INPUT = `${apiPrefixRC2}/account_book/:accountBookId/invoice/:invoiceId/input`,
+  DELETE_INVOICE_RC2_INPUT = `${apiPrefixRC2}/account_book/:accountBookId/invoice/input`,
   UPDATE_INVOICE_RC2_INPUT = `${apiPrefixRC2}/account_book/:accountBookId/invoice/:invoiceId/input`,
   CREATE_INVOICE_RC2_OUTPUT = `${apiPrefixRC2}/account_book/:accountBookId/invoice/output`,
   GET_INVOICE_RC2_OUTPUT = `${apiPrefixRC2}/account_book/:accountBookId/invoice/:invoiceId/output`,
-  DELETE_INVOICE_RC2_OUTPUT = `${apiPrefixRC2}/account_book/:accountBookId/invoice/:invoiceId/output`,
+  DELETE_INVOICE_RC2_OUTPUT = `${apiPrefixRC2}/account_book/:accountBookId/invoice/output`,
   UPDATE_INVOICE_RC2_OUTPUT = `${apiPrefixRC2}/account_book/:accountBookId/invoice/:invoiceId/output`,
   LIST_INVOICE_RC2_OUTPUT = `${apiPrefixRC2}/account_book/:accountBookId/invoice/output`,
   CERTIFICATE_GET_V2 = `${apiPrefixV2}/account_book/:accountBookId/certificate/:certificateId`,
