+++ src/pages/api/rc2/account_book/[accountBookId]/invoice/[invoiceId]/input/index.ts
@@ -12,11 +12,7 @@ import {
   logUserAction,
 } from '@/lib/utils/middleware';
 import { validateOutputData } from '@/lib/utils/validator';
-import {
-  findInvoiceRC2ById,
-  deleteInvoiceRC2,
-  updateInvoiceRC2Input,
-} from '@/lib/utils/repo/invoice_rc2.repo';
+import { findInvoiceRC2ById, updateInvoiceRC2Input } from '@/lib/utils/repo/invoice_rc2.repo';
 
 const handleGetRequest = async (req: NextApiRequest) => {
   const session = await getSession(req);
@@ -85,35 +81,6 @@ const handlePutRequest = async (req: NextApiRequest) => {
   return { response, statusMessage };
 };
 
-const handleDeleteRequest = async (req: NextApiRequest) => {
-  const session = await getSession(req);
-  let statusMessage: string = STATUS_MESSAGE.BAD_REQUEST;
-  let payload = null;
-
-  await checkSessionUser(session, APIName.DELETE_INVOICE_RC2_INPUT, req);
-  await checkUserAuthorization(APIName.DELETE_INVOICE_RC2_INPUT, req, session);
-
-  const { query, body } = checkRequestData(APIName.DELETE_INVOICE_RC2_INPUT, req, session);
-  if (!query || !body) throw new Error(STATUS_MESSAGE.INVALID_INPUT_PARAMETER);
-
-  const invoiceList = await deleteInvoiceRC2(session.userId, query.accountBookId, body.invoiceIds);
-
-  const { isOutputDataValid, outputData } = validateOutputData(
-    APIName.DELETE_INVOICE_RC2_INPUT,
-    invoiceList
-  );
-
-  if (!isOutputDataValid) {
-    statusMessage = STATUS_MESSAGE.INVALID_OUTPUT_DATA;
-  } else {
-    payload = outputData;
-    statusMessage = STATUS_MESSAGE.SUCCESS;
-  }
-
-  const response = formatApiResponse(statusMessage, payload);
-  return { response, statusMessage };
-};
-
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   const method = req.method || HttpMethod.GET;
   let httpCode = HTTP_STATUS.INTERNAL_SERVER_ERROR;
@@ -132,10 +99,7 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
         ({ response, statusMessage } = await handlePutRequest(req));
         ({ httpCode, result } = response);
         break;
-      case HttpMethod.DELETE:
-        ({ response, statusMessage } = await handleDeleteRequest(req));
-        ({ httpCode, result } = response);
-        break;
+
       default:
         statusMessage = STATUS_MESSAGE.METHOD_NOT_ALLOWED;
         ({ httpCode, result } = formatApiResponse<null>(statusMessage, null));
